requirements.txt
# Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.18
pydantic==2.10.3

# Camera & Inference
opencv-python-headless==4.10.0.84
numpy==1.26.4
onnxruntime==1.20.1
Pillow==11.0.0

# Frontend
streamlit==1.41.1
requests==2.32.3
